feat(api): extract service metadata parser, add per-op permissions, scaffold @carbon/api
Move the MCP generator's parse core into scripts/lib/service-metadata.ts
(pure buildAllToolMetadata()), and derive a per-operation permission
({module, actions}) from the service module + verb — validated against the
real route permissions (items->parts, account/shared->null). generate-mcp.ts
is now a thin writer; tool-metadata.json is byte-identical apart from the new
permission field (verified field-by-field over all 1495 ops). Scaffold the
internal @carbon/api package (manifest types + the jsonSchema Standard-Schema
wrapper and OpenAPI converter) and pin @orpc/* at ^1.15.0.

Note: the full build-time manifest relocation (decision 7) is deferred to
Docs Phase 2 — tool-metadata.json is also read by the workflow catalog
generator, a wider blast radius than the plan assumed.
