+++ apps/erp/package.json
@@ -12,6 +12,7 @@
     "@ai-sdk/openai": "catalog:",
     "@ai-sdk/react": "2.0.174",
     "@base-ui-components/react": "1.0.0-rc.0",
+    "@carbon/api": "workspace:*",
     "@carbon/auth": "workspace:*",
     "@carbon/database": "workspace:*",
     "@carbon/documents": "workspace:*",
@@ -46,6 +47,8 @@
     "@modelcontextprotocol/sdk": "1.29.0",
     "@monaco-editor/react": "4.7.0",
     "@nanostores/react": "catalog:",
+    "@orpc/openapi": "catalog:",
+    "@orpc/server": "catalog:",
     "@paper-design/shaders-react": "^0.0.57",
     "@react-aria/i18n": "catalog:",
     "@react-email/components": "catalog:",
