+++ packages/api/src/index.ts
@@ -1,0 +1,8 @@
+// @carbon/api — the Carbon API contract layer.
+//
+// The operation manifest (`@carbon/api/manifest`) is generated at build time; its
+// types live in `./manifest-types`. The oRPC contract and Standard-Schema wrapper
+// are added in `./contract` and `./schema`.
+
+export * from "./manifest-types";
+export * from "./schema";
