+++ packages/api/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@carbon/config/tsconfig/base.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
