+++ pnpm-workspace.yaml
@@ -69,6 +69,11 @@ catalog:
   '@logtape/redaction': 2.2.3
   '@logtape/testing': 2.2.3
   '@nanostores/react': 0.7.3
+  '@orpc/client': ^1.15.0
+  '@orpc/contract': ^1.15.0
+  '@orpc/openapi': ^1.15.0
+  '@orpc/openapi-client': ^1.15.0
+  '@orpc/server': ^1.15.0
   '@paralleldrive/cuid2': 2.2.2
   '@react-aria/i18n': 3.12.16
   '@react-email/components': 0.3.3
